+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tool/init/ItemInitializer.java
@@ -1,9 +1,14 @@
 package com.codetaylor.mc.pyrotech.modules.tool.init;
 
+import com.codetaylor.mc.athenaeum.reference.EnumMaterial;
+import com.codetaylor.mc.athenaeum.reference.ModuleMaterials;
 import com.codetaylor.mc.athenaeum.registry.Registry;
 import com.codetaylor.mc.athenaeum.util.ModelRegistrationHelper;
 import com.codetaylor.mc.pyrotech.modules.tool.ModuleTool;
 import com.codetaylor.mc.pyrotech.modules.tool.item.*;
+import com.codetaylor.mc.pyrotech.modules.tool.item.spi.*;
+import com.google.common.base.Preconditions;
+import net.minecraft.item.Item;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
@@ -12,38 +17,51 @@ public final class ItemInitializer {
 
   public static void onRegister(Registry registry) {
 
+    // crude
     registry.registerItem(new ItemCrudeAxe(), ItemCrudeAxe.NAME);
     registry.registerItem(new ItemCrudeHoe(), ItemCrudeHoe.NAME);
     registry.registerItem(new ItemCrudePickaxe(), ItemCrudePickaxe.NAME);
     registry.registerItem(new ItemCrudeShovel(), ItemCrudeShovel.NAME);
     registry.registerItem(new ItemCrudeFishingRod(), ItemCrudeFishingRod.NAME);
 
-    registry.registerItem(new ItemBoneAxe(), ItemBoneAxe.NAME);
-    registry.registerItem(new ItemBoneHoe(), ItemBoneHoe.NAME);
-    registry.registerItem(new ItemBonePickaxe(), ItemBonePickaxe.NAME);
-    registry.registerItem(new ItemBoneShovel(), ItemBoneShovel.NAME);
-    registry.registerItem(new ItemBoneSword(), ItemBoneSword.NAME);
+    // bone
+    {
+      Item.ToolMaterial material = EnumMaterial.BONE.getToolMaterial();
+      registry.registerItem(new ItemAxeBase(material, 8f, -3.2f, "bone"), ItemAxeBase.NAME_BONE);
+      registry.registerItem(new ItemHoeBase(material, "bone"), ItemHoeBase.NAME_BONE);
+      registry.registerItem(new ItemPickaxeBase(material, "bone"), ItemPickaxeBase.NAME_BONE);
+      registry.registerItem(new ItemShovelBase(material, "bone"), ItemShovelBase.NAME_BONE);
+      registry.registerItem(new ItemSwordBase(material, "bone"), ItemSwordBase.NAME_BONE);
+    }
 
-    registry.registerItem(new ItemFlintAxe(), ItemFlintAxe.NAME);
-    registry.registerItem(new ItemFlintHoe(), ItemFlintHoe.NAME);
-    registry.registerItem(new ItemFlintPickaxe(), ItemFlintPickaxe.NAME);
-    registry.registerItem(new ItemFlintShovel(), ItemFlintShovel.NAME);
-    registry.registerItem(new ItemFlintSword(), ItemFlintSword.NAME);
+    // flint
+    {
+      Item.ToolMaterial material = EnumMaterial.FLINT.getToolMaterial();
+      registry.registerItem(new ItemAxeBase(material, 8f, -3.2f, "flint"), ItemAxeBase.NAME_FLINT);
+      registry.registerItem(new ItemHoeBase(material, "flint"), ItemHoeBase.NAME_FLINT);
+      registry.registerItem(new ItemPickaxeBase(material, "flint"), ItemPickaxeBase.NAME_FLINT);
+      registry.registerItem(new ItemShovelBase(material, "flint"), ItemShovelBase.NAME_FLINT);
+      registry.registerItem(new ItemSwordBase(material, "flint"), ItemSwordBase.NAME_FLINT);
+    }
 
-    registry.registerItem(new ItemObsidianAxe(), ItemObsidianAxe.NAME);
-    registry.registerItem(new ItemObsidianHoe(), ItemObsidianHoe.NAME);
-    registry.registerItem(new ItemObsidianPickaxe(), ItemObsidianPickaxe.NAME);
-    registry.registerItem(new ItemObsidianShovel(), ItemObsidianShovel.NAME);
-    registry.registerItem(new ItemObsidianSword(), ItemObsidianSword.NAME);
+    // obsidian
+    {
+      Item.ToolMaterial material = Preconditions.checkNotNull(ModuleMaterials.OBSIDIAN);
+      registry.registerItem(new ItemAxeBase(material, 8f, -3.0f, "obsidian"), ItemAxeBase.NAME_OBSIDIAN);
+      registry.registerItem(new ItemHoeBase(material, "obsidian"), ItemHoeBase.NAME_OBSIDIAN);
+      registry.registerItem(new ItemPickaxeBase(material, "obsidian"), ItemPickaxeBase.NAME_OBSIDIAN);
+      registry.registerItem(new ItemShovelBase(material, "obsidian"), ItemShovelBase.NAME_OBSIDIAN);
+      registry.registerItem(new ItemSwordBase(material, "obsidian"), ItemSwordBase.NAME_OBSIDIAN);
+    }
 
     registry.registerItem(new ItemUnfiredClayShears(), ItemUnfiredClayShears.NAME);
-    registry.registerItem(new ItemClayShears(), ItemClayShears.NAME);
-    registry.registerItem(new ItemStoneShears(), ItemStoneShears.NAME);
-    registry.registerItem(new ItemBoneShears(), ItemBoneShears.NAME);
-    registry.registerItem(new ItemFlintShears(), ItemFlintShears.NAME);
+    registry.registerItem(new ItemShearsBase("clay"), ItemShearsBase.NAME_CLAY);
+    registry.registerItem(new ItemShearsBase("stone"), ItemShearsBase.NAME_STONE);
+    registry.registerItem(new ItemShearsBase("bone"), ItemShearsBase.NAME_BONE);
+    registry.registerItem(new ItemShearsBase("flint"), ItemShearsBase.NAME_FLINT);
     registry.registerItem(new ItemGoldShears(), ItemGoldShears.NAME);
-    registry.registerItem(new ItemDiamondShears(), ItemDiamondShears.NAME);
-    registry.registerItem(new ItemObsidianShears(), ItemObsidianShears.NAME);
+    registry.registerItem(new ItemShearsBase("diamond"), ItemShearsBase.NAME_DIAMOND);
+    registry.registerItem(new ItemShearsBase("obsidian"), ItemShearsBase.NAME_OBSIDIAN);
   }
 
   @SideOnly(Side.CLIENT)