+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tool/item/spi/IFlintTierTool.java
@@ -1,0 +1,12 @@
+package com.codetaylor.mc.pyrotech.modules.tool.item.spi;
+
+public interface IFlintTierTool
+    extends ITierTool {
+
+  @Override
+  default String getToolTierName() {
+
+    return "flint";
+  }
+
+}