+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tool/item/spi/ITierTool.java
@@ -1,0 +1,6 @@
+package com.codetaylor.mc.pyrotech.modules.tool.item.spi;
+
+public interface ITierTool {
+
+  String getToolTierName();
+}